Add: FAT16 library to communicate with SD Card Reader; Update: SBSBS to swap large memory using SD library for smaller FAT16 library

+++ ArduinoCode/libraries/Queuetue_HX711_Library/src/Q2HX711.h
@@ -12,7 +12,7 @@
 // const uint16_t ADJUST_AVERAGE = 500;
 const long HX711_WEIGHT_BOUNDARY_TRIGGER = 17800l; // Difference in weight to "detect" snake, 17800 / 712 = 25 grams
 const unsigned short INITIAL_WEIGHT_POLLING = 1000;
-const unsigned short FAST_WEIGHT_POLLING = 100;
+const byte FAST_WEIGHT_POLLING = 100;
 
 class Q2HX711
 {
